+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2022-11-22 22:09:26
+
+- Improve the SubPicker by allowing extra choices to be sent back to the main page
+
 ## 2022-11-21 22:13:38
 
 - Add support to the `SubPicker` to show the chooser menu and active options